+++ examples/data-sources/okta_identity_source_sessions/datasource.tf
@@ -1,3 +1,4 @@
 data "okta_identity_source_sessions" "test" {
   identity_source_id = "0oaxc95befZNgrJl71d7"
+  session_id = "apsydhsafnSPHlZyG1d7"
 }